a2/consumer/requirements.txt
kafka-python
pymongo
fastapi[standard]
